Complete folder and file restructure
