+++ config/webcams/alps.js
