+++ config/webcams/clyde.js
