+++ config/webcams/fiords.js
