+++ config/webcams/gore.js
