+++ config/webcams/windward.js
